remote-agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pillow>=10.0.0
pyautogui>=0.9.54
mss>=9.0.0
pyperclip>=1.8.0
